+++ src/main/java/xyz/wavey/rentalservice/service/InsuranceService.java
@@ -1,7 +1,7 @@
 package xyz.wavey.rentalservice.service;
 
 import xyz.wavey.rentalservice.model.Insurance;
-import xyz.wavey.rentalservice.vo.RequestAddInsurance;
+import xyz.wavey.rentalservice.vo.request.RequestAddInsurance;
 
 import java.util.List;
