+++ src/main/java/xyz/wavey/rentalservice/service/KakaoPayOpenFeign.java
@@ -4,7 +4,10 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.cloud.openfeign.SpringQueryMap;
 import org.springframework.web.bind.annotation.PostMapping;
 import xyz.wavey.rentalservice.base.KakaoOpenFeignConfig;
-import xyz.wavey.rentalservice.vo.*;
+import xyz.wavey.rentalservice.vo.request.KakaoPayApproveParameter;
+import xyz.wavey.rentalservice.vo.request.KakaoPayReadyParameter;
+import xyz.wavey.rentalservice.vo.response.ResponseKakaoPayApprove;
+import xyz.wavey.rentalservice.vo.response.ResponseKakaoPayReady;
 
 @FeignClient(name = "kakaopay-service", url = "https://kapi.kakao.com", configuration = KakaoOpenFeignConfig.class)
 public interface KakaoPayOpenFeign {
