+++ src/main/java/xyz/wavey/rentalservice/service/PurchaseService.java
@@ -1,9 +1,9 @@
 package xyz.wavey.rentalservice.service;
 
-import xyz.wavey.rentalservice.vo.RequestAddRental;
-import xyz.wavey.rentalservice.vo.RequestKakaoPayApprove;
-import xyz.wavey.rentalservice.vo.ResponseKakaoPayApprove;
-import xyz.wavey.rentalservice.vo.ResponseKakaoPayReady;
+import xyz.wavey.rentalservice.vo.request.RequestAddRental;
+import xyz.wavey.rentalservice.vo.request.RequestKakaoPayApprove;
+import xyz.wavey.rentalservice.vo.response.ResponseKakaoPayApprove;
+import xyz.wavey.rentalservice.vo.response.ResponseKakaoPayReady;
 
 public interface PurchaseService {
