+++ src/main/java/xyz/wavey/rentalservice/vo/request/KakaoPayApproveParameter.java
@@ -1,4 +1,4 @@
-package xyz.wavey.rentalservice.vo;
+package xyz.wavey.rentalservice.vo.request;
 
 import lombok.Builder;
 import lombok.Data;
